+++ src/Enum/WirelessOperatingMode.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\Enum;
+
+enum WirelessOperatingMode: string
+{
+    case Adhoc = 'A';
+    case Infrastructure = 'I';
+}